resume-ai-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
PyMuPDF           # for PDF text extraction
python-dotenv     # optional (if using .env for API keys)
